+++ src/dm_preprocessor/directive/ifdef.rs
@@ -1,9 +1,6 @@
 use log::warn;
 
-use crate::{
-    dm_preprocessor::{token_handling::DmToken, DmPreProcessor},
-    util::ParseError,
-};
+use crate::{dm_preprocessor::lib::DmPreProcessor, tokens::dm_token::DmToken, util::ParseError};
 
 impl DmPreProcessor {
     pub(super) fn handle_ifdef(&mut self, args: &[DmToken]) -> Result<(), ParseError> {
