+++ src/dm_preprocessor/directive/include.rs
@@ -1,9 +1,6 @@
 use log::{error, trace};
 
-use crate::{
-    dm_preprocessor::{token_handling::DmToken, DmPreProcessor},
-    util::ParseError,
-};
+use crate::{dm_preprocessor::lib::DmPreProcessor, tokens::dm_token::DmToken, util::ParseError};
 
 impl DmPreProcessor {
     pub(super) fn handle_include(&mut self, args: &[DmToken]) -> Result<(), ParseError> {
