+++ src/dm_preprocessor/tests/mod.rs
@@ -1,1 +1,1 @@
-mod token_handling;
+mod lib;
