+++ src/tokens/constants.rs
@@ -1,0 +1,3 @@
+pub const SYM_MATH: &[char; 8] = &['+', '-', '*', '/', '%', '^', '&', '|'];
+pub const SYM_QUTE: &[char; 2] = &['"', '\''];
+pub const SYM_OTHR: &[char; 6] = &['!', '=', '.', '#', '<', '>'];
