+++ src/tokens/determine_token_action.rs
@@ -1,42 +1,15 @@
-use std::{fmt::Display, iter::Peekable, str::Chars};
+use std::{iter::Peekable, str::Chars};
 
-use crate::dm_preprocessor::token_handling::{DmToken, TokenizeState};
-
-use super::{
-    count_backslashes, start_new_token::get_default_token_action,
-    whitespace_char::is_first_non_whitespace_char,
+use crate::{
+    dm_preprocessor::tokenize_state::TokenizeState,
+    util::{count_backslashes, whitespace_char::is_first_non_whitespace_char},
 };
 
-#[derive(Debug, PartialEq)]
-pub enum TokenAction {
-    /// Start a new token with the current character.
-    StartNewToken,
-    /// Add the character to the current token.
-    ContinueToken,
-    /// Add the character to the current token and end the token.
-    EndToken,
-    /// Isolate the current character into its own token.
-    IsolateToken,
-    /// Ignore the current character.
-    None,
-    /// Drop the current token and character.
-    DropToken,
-}
+use super::{
+    default_token_action::get_default_token_action, dm_token::DmToken, token_action::TokenAction,
+};
 
-impl Display for TokenAction {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TokenAction::StartNewToken => write!(f, "StartNewToken"),
-            TokenAction::ContinueToken => write!(f, "ContinueToken"),
-            TokenAction::EndToken => write!(f, "EndToken"),
-            TokenAction::IsolateToken => write!(f, "IsolateToken"),
-            TokenAction::None => write!(f, "None"),
-            TokenAction::DropToken => write!(f, "DropToken"),
-        }
-    }
-}
-
-// Determines what action should be taken for the current character.
+// Non default token actions
 pub fn determine_token_action(
     state: &mut TokenizeState,
     char: char,
