+++ src/tokens/tests/comment.rs
@@ -1,10 +1,9 @@
 #[test]
-fn test_tokenize_multiple_lines() {
-    use crate::dm_preprocessor::{DmPreProcessor, DmToken};
+fn test_tokenize_comment() {
+    use crate::{dm_preprocessor::lib::DmPreProcessor, tokens::dm_token::DmToken};
 
     let mut preprocessor = DmPreProcessor::new();
-
-    let lines = vec!["This is a test.", "Another test."];
+    let lines = vec!["This is a test.", "// This is a comment.", "Another test."];
 
     let expected = vec![
         DmToken::from("This"),
@@ -16,6 +15,17 @@ fn test_tokenize_multiple_lines() {
         DmToken::from("test"),
         DmToken::from("."),
         DmToken::from("\n"),
+        DmToken::from("//"),
+        DmToken::from(" "),
+        DmToken::from("This"),
+        DmToken::from(" "),
+        DmToken::from("is"),
+        DmToken::from(" "),
+        DmToken::from("a"),
+        DmToken::from(" "),
+        DmToken::from("comment"),
+        DmToken::from("."),
+        DmToken::from("\n"),
         DmToken::from("Another"),
         DmToken::from(" "),
         DmToken::from("test"),
@@ -24,6 +34,5 @@ fn test_tokenize_multiple_lines() {
     ];
 
     let result = preprocessor.tokenize(&lines);
-
     assert_eq!(result, expected);
 }
