+++ src/tokens/tests/condense.rs
@@ -1,9 +1,10 @@
 #[test]
-fn test_tokenize_comment() {
-    use crate::dm_preprocessor::{DmPreProcessor, DmToken};
+fn test_condense_lines() {
+    use crate::{dm_preprocessor::lib::DmPreProcessor, tokens::dm_token::DmToken};
 
     let mut preprocessor = DmPreProcessor::new();
-    let lines = vec!["This is a test.", "// This is a comment.", "Another test."];
+
+    let lines = vec!["This is a test. \\", "Another test."];
 
     let expected = vec![
         DmToken::from("This"),
@@ -14,18 +15,7 @@ fn test_tokenize_comment() {
         DmToken::from(" "),
         DmToken::from("test"),
         DmToken::from("."),
-        DmToken::from("\n"),
-        DmToken::from("//"),
         DmToken::from(" "),
-        DmToken::from("This"),
-        DmToken::from(" "),
-        DmToken::from("is"),
-        DmToken::from(" "),
-        DmToken::from("a"),
-        DmToken::from(" "),
-        DmToken::from("comment"),
-        DmToken::from("."),
-        DmToken::from("\n"),
         DmToken::from("Another"),
         DmToken::from(" "),
         DmToken::from("test"),
@@ -34,5 +24,6 @@ fn test_tokenize_comment() {
     ];
 
     let result = preprocessor.tokenize(&lines);
+
     assert_eq!(result, expected);
 }
