+++ src/util/whitespace_char.rs
@@ -1,4 +1,4 @@
-use crate::dm_preprocessor::token_handling::DmToken;
+use crate::tokens::dm_token::DmToken;
 
 // Returns true if the line is empty or only contains whitespace.
 pub fn is_first_non_whitespace_char(line_tokens: &[DmToken]) -> bool {
@@ -10,6 +10,8 @@ pub fn is_first_non_whitespace_char(line_tokens: &[DmToken]) -> bool {
 
 #[cfg(test)]
 mod tests {
+    use crate::tokens::dm_token::DmToken;
+
     use super::*;
 
     #[test]
